refactor: use create_frames() from evaluate.py in CLI frames commands
