+++ src/wildcamtools/cli/frames.py
@@ -1,19 +1,13 @@
-import logging
 from pathlib import Path
 from typing import Annotated
 
 import typer
 
-from wildcamtools.lib import Frame, FrameHandler
 from wildcamtools.lib.errors.cli import OutputNotDirectoryError
-from wildcamtools.lib.frames import CropPanHandler, FilterSSIM, FrameImageWriter, Rescaler
-from wildcamtools.lib.motion import FlowMotion
-from wildcamtools.lib.stats import get_video_stats
+from wildcamtools.lib.frames import AIFrameCreation, create_frames
 from wildcamtools.lib.timing import Timer
-from wildcamtools.lib.vidio import VideoReader
 
 app = typer.Typer()
-logger = logging.getLogger(__name__)
 
 
 @app.command()
@@ -25,32 +19,28 @@ def ssim(
     y: int | None = None,
     fps: float | None = None,
 ) -> None:
-
     if output is None:
         output = input_.with_suffix("")
 
     if output.exists() and not output.is_dir():
         raise OutputNotDirectoryError()
 
-    stats = get_video_stats(input_)
-
-    handlers: list[FrameHandler] = []
-    if x or y or fps:
-        handlers.append(Rescaler(stats=stats, x=x, y=y, fps=fps))
-    if threshold:
-        handlers.append(FilterSSIM(threshold))
-    handlers.append(FrameImageWriter(output))
+    frame_creation = AIFrameCreation(
+        filename=Path(input_.name),
+        video_directory=input_.parent,
+        tmpdir=output,
+        x=x,
+        y=y,
+        fps=fps,
+        similarity_minimum=threshold,
+    )
 
     timer = Timer()
+    with timer:
+        result = create_frames(frame_creation)
 
-    with VideoReader(input_) as video_input:
-        frame: Frame
-        for frame in video_input:
-            with timer:
-                for handler in handlers:
-                    frame = handler.handle(frame)
-
-    typer.secho(f"Processed {timer.intervals:d} frames in {timer.elapsed:.2f} sec; {timer.per_second:.2f}FPS")
+    fps_actual = result.frame_count / timer.elapsed if timer.elapsed > 0 else 0.0
+    typer.secho(f"Processed {result.frame_count:d} frames in {timer.elapsed:.2f} sec; {fps_actual:.2f}FPS")
 
 
 @app.command()
@@ -67,47 +57,34 @@ def flow(
     crop_inertia: float = 10.0,
     similarity_minimum: float = 0.5,
 ) -> None:
-
     if output is None:
         output = input_.with_suffix("")
 
     if output.exists() and not output.is_dir():
         raise OutputNotDirectoryError()
-    output.mkdir(parents=True, exist_ok=True)
 
-    stats = get_video_stats(input_)
+    frame_creation = AIFrameCreation(
+        filename=Path(input_.name),
+        video_directory=input_.parent,
+        tmpdir=output,
+        history=history,
+        threshold=threshold,
+        kernel_size=kernel_size,
+        x=x,
+        y=y,
+        fps=fps,
+        crop_expansion=crop_expansion,
+        crop_inertia=crop_inertia,
+        similarity_minimum=similarity_minimum,
+        do_croppan=True,
+        motion_type="flow",
+    )
     timer = Timer()
+    with timer:
+        result = create_frames(frame_creation)
 
-    handlers: list[FrameHandler] = []
-    handlers.append(
-        CropPanHandler(
-            motion_handler=FlowMotion(
-                history=history,
-                threshold=threshold,
-                kernel_size=kernel_size,
-            ),
-            expansion=crop_expansion,
-            inertia=crop_inertia,
-        )
-    )
-    handlers.append(FilterSSIM(similarity_minimum=similarity_minimum))
-    handlers.append(
-        Rescaler(
-            stats=stats,
-            x=x,
-            y=y,
-            fps=fps,
-        )
-    )
-    handlers.append(FrameImageWriter(output))
-
-    with VideoReader(input_) as video_input:
-        for frame in video_input:
-            with timer:
-                for handler in handlers:
-                    frame = handler.handle(frame)
-
-    typer.secho(f"Processed {timer.intervals:d} frames in {timer.elapsed:.2f} sec; {timer.per_second:.2f}FPS")
+    fps_actual = result.frame_count / timer.elapsed if timer.elapsed > 0 else 0.0
+    typer.secho(f"Processed {result.frame_count:d} frames in {timer.elapsed:.2f} sec; {fps_actual:.2f}FPS")
 
 
 if __name__ == "__main__":
